release 0.1.216 source code

+++ services/codeartspipeline/v2/model/model_list_pipeline_templates_response.go
@@ -10,13 +10,13 @@ import (
 type ListPipelineTemplatesResponse struct {
 
 	// **参数解释**： 起始偏移。 **取值范围**： 不涉及。
-	Offset *int32 `json:"offset,omitempty"`
+	Offset *int64 `json:"offset,omitempty"`
 
 	// **参数解释**： 查询大小。 **取值范围**： 不涉及。
-	Limit *int32 `json:"limit,omitempty"`
+	Limit *int64 `json:"limit,omitempty"`
 
 	// **参数解释**： 记录总数。 **取值范围**： 不涉及。
-	Total *int32 `json:"total,omitempty"`
+	Total *int64 `json:"total,omitempty"`
 
 	// **参数解释**： 流水线模板列表，包含流水线模板的详细信息。 **取值范围**： 不涉及。
 	Templates      *[]PipelineTemplateSimpleVo `json:"templates,omitempty"`
